pyproject.toml
[project]
name = "gerenciador-estoque"
version = "0.1.0"
description = ""
authors = [
    {name = "luanfred",email = "luanmatheussf4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "python-jose (>=3.4.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "app"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.9"
taskipy = "^1.14.1"
ignr = "^2.2"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'uvicorn app.main:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'
create_tables = 'python3 -m app.create_all_tables'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'